API操作の関数とDOM操作の関数を別ファイルに分ける
変更内容：API操作の関数をapi.jsに、DOM操作の関数をdom.jsに分けた。
