+++ index.html
@@ -48,6 +48,6 @@
       <p>© 2024 タスク管理アプリ</p>
     </footer>
     <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.21/lodash.min.js"></script>
-    <script src="./task_add.js" type="module"></script>
+    <script src="./dom.js" type="module"></script>
   </body>
 </html>
